+++ KraftonEngine/Source/Editor/UI/Asset/StaticMeshEditorWidget.cpp
@@ -94,7 +94,7 @@ void FStaticMeshEditorWidget::Open(UObject* Object)
 
 	ImVec2 ViewportSize = ImGui::GetContentRegionAvail();
 
-	ViewportClient.Initialize(GEngine->GetRenderer().GetFD3DDevice().GetDevice(), ViewportSize.x, ViewportSize.y);
+	ViewportClient.Initialize(GEngine->GetRenderer().GetFD3DDevice().GetDevice(), static_cast<uint32>(ViewportSize.x), static_cast<uint32>(ViewportSize.y));
 	ViewportClient.SetPreviewWorld(WorldContext.World);
 	ViewportClient.SetPreviewActor(Actor);
 	ViewportClient.SetPreviewMeshComponent(Actor->GetComponentByClass<UStaticMeshComponent>());
