+++ KraftonEngine/Source/Engine/Render/Shadow/ShadowAtlasQuadTree.cpp
@@ -69,7 +69,7 @@ TArray<FAtlasRegion> FShadowAtlasQuadTree::CommitBatch() {
 	for (int32 OrigIdx : Order) {
 		auto desired_res = static_cast<float>(RoundToNearestPowerOfTwo(static_cast<uint32>(Batch[OrigIdx].Resolution)));
 		desired_res = desired_res > MinShadowMapResolution ? desired_res : MinShadowMapResolution;
-		FAtlasRegion AtlasRegion = AllocateNode(0, desired_res, Batch[OrigIdx].LightIdx);
+		FAtlasRegion AtlasRegion = AllocateNode(0, static_cast<uint32>(desired_res), Batch[OrigIdx].LightIdx);
 		Results[OrigIdx] = AtlasRegion;
 	}
